git-svn-id: http://springhead.info/spr2/Springhead2/trunk@2880 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ include/Creature/SprCRBody.h
@@ -500,7 +500,7 @@ struct CRTrunkFootAnimalBodyDesc : CRBodyDesc{
 	double waistHeight, waistBreadth, waistThickness;
 	double chestHeight, chestBreadth, chestThickness;
 	double neckLength;
-	double headDiameter;
+	float  headDiameter;
 	double footLength, footBreadth, footThickness;
 
 	/// 各関節のバネダンパ
